update code with feedback
