+++ gradle.properties
@@ -14,7 +14,7 @@ org.gradle.jvmargs=-Xmx1536m
 android.debug.obsoleteApi=true
 
 ## PUBLISHING VARS
-VERSION_NAME=0.0.4.1-SNAPSHOT
+VERSION_NAME=0.0.5-SNAPSHOT
 VERSION_CODE=1
 GROUP=org.smartregister
 POM_SETTING_DESCRIPTION=OpenSRP Client Maternity Library
